+++ lib/bunnycdn/engine.rb
@@ -15,20 +15,5 @@ module Bunnycdn
         prepend Bunnycdn::ViewOverrides if Bunnycdn.configuration.enhance_image_tag
       end
     end
-
-    initializer "bunnycdn.configure_from_rails" do |app|
-      config = Bunnycdn.configuration
-
-      if app.config.respond_to?(:x) && app.config.x.respond_to?(:bunnycdn)
-        bunny = app.config.x.bunnycdn
-
-        config.uploads_zone_url ||= bunny.respond_to?(:uploads_zone_url) ? bunny.uploads_zone_url : nil
-        config.uploads_prefix ||= bunny.respond_to?(:uploads_prefix) ? bunny.uploads_prefix : nil
-        config.static_zone_url ||= bunny.respond_to?(:static_zone_url) ? bunny.static_zone_url : nil
-        config.enhance_image_tag = bunny.enhance_image_tag if bunny.respond_to?(:enhance_image_tag)
-        config.default_quality = bunny.default_quality if bunny.respond_to?(:default_quality)
-        config.default_format = bunny.default_format if bunny.respond_to?(:default_format)
-      end
-    end
   end
 end